+++ src/main/java/com/code/alphavn/controller/adminController/ServletEcommerceProduct.java
@@ -4,6 +4,7 @@ import com.code.alphavn.model.Order;
 import com.code.alphavn.model.Product;
 import com.code.alphavn.model.ProductDiscount;
 import com.code.alphavn.model.ProductInfo;
+import com.code.alphavn.model.adminModel.Admin;
 import com.code.alphavn.service.adminService.AdminServiceImpl;
 import com.code.alphavn.service.userService.UserServiceImpl;
 
@@ -14,6 +15,7 @@ import java.io.*;
 import java.nio.file.Paths;
 import java.sql.Date;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 
 @WebServlet(name = "ServletEcommerceProduct", value = "/ecommerce-product")
@@ -95,6 +97,9 @@ public class ServletEcommerceProduct extends HttpServlet {
                 case "product-details":
                     updateProduct(request, response);
                     break;
+                case "deleteSelected":
+                    handleDeleteSelected(request, response);
+                    break;
                 default:
                     System.out.println("This is the do post Ecommerce product !");
                     break;
@@ -109,48 +114,76 @@ public class ServletEcommerceProduct extends HttpServlet {
     //    ================== HANDLE VIEW OF ECOMMERCE PRODUCT ====================
 
     public void showProductList(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        adminService = new AdminServiceImpl();
-        request.setAttribute("products", adminService.getAllProducts());
-        System.out.println(adminService.getAllProducts().size());
-        request.getRequestDispatcher("components/adminComponents/ecommerce-products.jsp").forward(request, response);
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            adminService = new AdminServiceImpl();
+            request.setAttribute("products", adminService.getAllProducts());
+            System.out.println(adminService.getAllProducts().size());
+            request.getRequestDispatcher("components/adminComponents/ecommerce-products.jsp").forward(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
 
     public void showProductDetails(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        int pid = Integer.parseInt(request.getParameter("id"));
-        request.setAttribute("product", adminService.getProductByID(pid));
-        request.setAttribute("discount", adminService.findDiscountProductById(pid));
-        request.getRequestDispatcher("components/adminComponents/ecommerce-product-details.jsp").forward(request, response);
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            int pid = Integer.parseInt(request.getParameter("id"));
+            request.setAttribute("product", adminService.getProductByID(pid));
+            request.setAttribute("discount", adminService.findDiscountProductById(pid));
+            request.getRequestDispatcher("components/adminComponents/ecommerce-product-details.jsp").forward(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
 
     public void showOverView(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        String week = request.getParameter("week");
-        System.out.println(week);
-        request.setAttribute("statistic", adminService.manageProducts(Integer.parseInt(week)));
-        request.setAttribute("statistic2", adminService.manageProducts(Integer.parseInt(week) - 1));
 
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            String week = request.getParameter("week");
+            System.out.println(week);
+            request.setAttribute("statistic", adminService.manageProducts(Integer.parseInt(week)));
+            request.setAttribute("statistic2", adminService.manageProducts(Integer.parseInt(week) - 1));
+            request.setAttribute("manages", adminService.getManageProductInCurrDate());
+            request.setAttribute("prdTops", adminService.getTopSellingProduct());
+            request.getRequestDispatcher("/components/adminComponents/ecommerce.jsp").forward(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
 
-        request.setAttribute("manages", adminService.getManageProductInCurrDate());
+        
 
-
-        request.setAttribute("prdTops", adminService.getTopSellingProduct());
-
-
-        request.getRequestDispatcher("/components/adminComponents/ecommerce.jsp").forward(request, response);
     }
 
 
     public void showAddProductLayOut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
         request.getRequestDispatcher("components/adminComponents/ecommerce-add-product.jsp").forward(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
 
     public void deleleProduct(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        int pid = Integer.parseInt(request.getParameter("pid"));
-        boolean isDeleted = adminService.deleteProduct(pid);
-        if (isDeleted) {
-            request.setAttribute("success", "Delete product successfully!");
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            int pid = Integer.parseInt(request.getParameter("pid"));
+            boolean isDeleted = adminService.deleteProduct(pid);
+            if (isDeleted) {
+                request.setAttribute("success", "Delete product successfully!");
+            }
+            request.setAttribute("error", "Delete product faild!");
+            showProductList(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
         }
-        request.setAttribute("error", "Delete product faild!");
-        showProductList(request, response);
     }
 
     //    ================== END HANDLE VIEW OF ECOMMERCE PRODUCT ====================
@@ -182,6 +215,10 @@ public class ServletEcommerceProduct extends HttpServlet {
     }
 
     public void createProduct(HttpServletRequest request, HttpServletResponse response) throws SQLException, ServletException, IOException {
+
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
         String prdName = request.getParameter("productName");
         String files = request.getParameter("fileNames");
         String prdDesc = request.getParameter("description");
@@ -210,57 +247,64 @@ public class ServletEcommerceProduct extends HttpServlet {
                     Date.valueOf(startDate), Date.valueOf(endDate));
         }
 
-        boolean isCreated = adminService.createProduct(productInfo, productDiscount);
+            boolean isCreated = adminService.createProduct(productInfo, productDiscount);
 
-        if (!isCreated) {
-            request.setAttribute("error", "Create new product faild, Try again !");
+            if (!isCreated) {
+                request.setAttribute("error", "Create new product faild, Try again !");
+            }
+            request.setAttribute("success", "Create product successfully !");
+            response.sendRedirect("ecommerce-product?action=products");
+        } else {
+            response.sendRedirect("loginCustomer");
         }
-        request.setAttribute("success", "Create product successfully !");
-        response.sendRedirect("ecommerce-product?action=products");
     }
 
 
     public void updateProduct(HttpServletRequest request, HttpServletResponse response) throws SQLException, ServletException, IOException {
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            String pid = request.getParameter("pid");
 
-        String pid = request.getParameter("pid");
+            String prdName = request.getParameter("productName");
+            String files = request.getParameter("fileNames");
+            String prdDesc = request.getParameter("description");
+            String price = request.getParameter("product-price");
+            String vendor = request.getParameter("vendor");
+            String category = request.getParameter("category");
+            String prdQuantity = request.getParameter("quantity");
+            String discount = request.getParameter("product-discount");
+            String[] splitFiles = files.split(" ; ");
 
-        String prdName = request.getParameter("productName");
-        String files = request.getParameter("fileNames");
-        String prdDesc = request.getParameter("description");
-        String price = request.getParameter("product-price");
-        String vendor = request.getParameter("vendor");
-        String category = request.getParameter("category");
-        String prdQuantity = request.getParameter("quantity");
-        String discount = request.getParameter("product-discount");
 
+            Product product = new Product(Integer.parseInt(pid), prdName, prdDesc, Integer.parseInt(prdQuantity), category);
 
-        String[] splitFiles = files.split(" ; ");
+            ProductInfo productInfo = new ProductInfo(product, Double.parseDouble(price), splitFiles[0], splitFiles[1], splitFiles[2], vendor);
 
-        Product product = new Product(Integer.parseInt(pid), prdName, prdDesc, Integer.parseInt(prdQuantity), category);
+            // == CHECK THE PRODUCT HAVE THE DISCOUNT OR NOT AND HANDLE
+            ProductDiscount productDiscount = null;
 
-        ProductInfo productInfo = new ProductInfo(product, Double.parseDouble(price), splitFiles[0], splitFiles[1], splitFiles[2], vendor);
+            if (!discount.equalsIgnoreCase("0")) {
+                String startDate = request.getParameter("start-date");
+                String endDate = request.getParameter("end-date");
+                String discountName = request.getParameter("discount-name");
+                productDiscount = new ProductDiscount(Integer.parseInt(pid), discountName,
+                        Double.parseDouble(discount), Date.valueOf(startDate), Date.valueOf(endDate));
+            } else {
+                adminService.deleteDiscountByPID(Integer.parseInt(pid));
+            }
 
-        // == CHECK THE PRODUCT HAVE THE DISCOUNT OR NOT AND HANDLE
-        ProductDiscount productDiscount = null;
+            // == HANDLE UPADTE PRODUCT
+            boolean isCreated = adminService.updateProduct(productInfo, productDiscount);
 
-        if (!discount.equalsIgnoreCase("0")) {
-            String startDate = request.getParameter("start-date");
-            String endDate = request.getParameter("end-date");
-            String discountName = request.getParameter("discount-name");
-            productDiscount = new ProductDiscount(Integer.parseInt(pid), discountName,
-                    Double.parseDouble(discount), Date.valueOf(startDate), Date.valueOf(endDate));
+            if (!isCreated) {
+                request.setAttribute("error", "Create new product faild, Try again !");
+            }
+            request.setAttribute("success", "Create product successfully !");
+            response.sendRedirect("ecommerce-product?action=products");
         } else {
-            adminService.deleteDiscountByPID(Integer.parseInt(pid));
+            response.sendRedirect("loginCustomer");
         }
-
-        // == HANDLE UPADTE PRODUCT
-        boolean isCreated = adminService.updateProduct(productInfo, productDiscount);
-
-        if (!isCreated) {
-            request.setAttribute("error", "Create new product faild, Try again !");
-        }
-        request.setAttribute("success", "Create product successfully !");
-        response.sendRedirect("ecommerce-product?action=products");
     }
 
     // =================== END HANDLE DO POST OF ECOMMERCE PRODUCT =====================
@@ -268,8 +312,8 @@ public class ServletEcommerceProduct extends HttpServlet {
     public void nextProductDetail(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, SQLException {
         UserServiceImpl userService = new UserServiceImpl();
         HttpSession session = request.getSession();
-//        Customer account = (Customer) session.getAttribute("acc");
-//        if (account != null) {
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
         int pid = Integer.parseInt(request.getParameter("id"));
         List<ProductInfo> products = adminService.getAllProducts();
         int currentIndex = getProductIndex(products, pid);
@@ -278,17 +322,17 @@ public class ServletEcommerceProduct extends HttpServlet {
 
         response.sendRedirect("ecommerce-product?action=product-details&id=" + nextProductId);
 
-//        } else {
-//            response.sendRedirect("loginCustomer");
-//        }
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
 
     }
 
     public void priviousProductDetail(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, SQLException {
         UserServiceImpl userService = new UserServiceImpl();
         HttpSession session = request.getSession();
-//        Customer account = (Customer) session.getAttribute("acc");
-//        if (account != null) {
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
         int pid = Integer.parseInt(request.getParameter("id"));
         List<ProductInfo> products = adminService.getAllProducts();
         int currentIndex = getProductIndex(products, pid);
@@ -296,9 +340,9 @@ public class ServletEcommerceProduct extends HttpServlet {
         int previousProductId = getPreviousProductId(products, currentIndex);
 
         response.sendRedirect("ecommerce-product?action=product-details&id=" + previousProductId);
-//        } else {
-//            response.sendRedirect("loginCustomer");
-//        }
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
 
     }
 
@@ -329,6 +373,29 @@ public class ServletEcommerceProduct extends HttpServlet {
             return nextProduct.getProduct().getId();
         }
         return productList.get(0).getProduct().getId(); // Không có product liền sau
+    }
+
+    public void handleDeleteSelected (HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            String orderIdsParam = request.getParameter("productIds");
+            String[] selectedIds = orderIdsParam.split(",");
+            List<Product> productIds = new ArrayList<>();
+            if (selectedIds != null) {
+                for (String id : selectedIds) {
+                    int Id = Integer.parseInt(id);
+                    productIds.add( new Product(Id));
+                }
+            }
+            System.out.println("orderIdsParam"+orderIdsParam);
+            System.out.println("selectedIds"+selectedIds);
+            System.out.println("productIds"+productIds);
+            adminService.DeleteProductSelected(productIds);
+            response.sendRedirect("ecommerce-product?action=products");
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
 }
