+++ src/main/java/com/code/alphavn/controller/adminController/ServletManageReviews.java
@@ -1,5 +1,6 @@
 package com.code.alphavn.controller.adminController;
 
+import com.code.alphavn.model.adminModel.Admin;
 import com.code.alphavn.service.adminService.AdminServiceImpl;
 
 import javax.servlet.ServletException;
@@ -7,6 +8,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.sql.SQLException;
 
@@ -43,16 +45,29 @@ public class ServletManageReviews extends HttpServlet {
     }
 
     public void ShowManageReviews(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        request.setAttribute("reviews", adminService.getProductReviews());
-        request.setAttribute("productRatings",adminService.getProductRatingReviews());
-        request.getRequestDispatcher("/components/adminComponents/manageReviews.jsp").forward(request, response);
-        System.out.println(adminService.getProductReviews());
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            request.setAttribute("reviews", adminService.getProductReviews());
+            request.setAttribute("productRatings",adminService.getProductRatingReviews());
+            request.getRequestDispatcher("/components/adminComponents/manageReviews.jsp").forward(request, response);
+            System.out.println(adminService.getProductReviews());
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
     public void DeleteReview(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException {
-        String id= (request.getParameter("id"));
-        System.out.println(id);
+        HttpSession session = request.getSession();
+        Admin account = (Admin) session.getAttribute("acc");
+        if (account != null) {
+            String id= (request.getParameter("id"));
+            System.out.println(id);
 
-       adminService.deleteReviewById(Integer.parseInt(id));
-        ShowManageReviews(request, response);
+            adminService.deleteReviewById(Integer.parseInt(id));
+            ShowManageReviews(request, response);
+        } else {
+            response.sendRedirect("loginCustomer");
+        }
     }
+
 }
